+++ accounts/urls.py
@@ -6,4 +6,5 @@ urlpatterns = [
     path("login/", views.LoginView.as_view()),
     path("logout/", views.LogoutView.as_view()),
     path("withdraw/", views.WithdrawView.as_view()),
+    path("password/", views.PasswordChangeView.as_view()),
 ]
