+++ users/views.py
@@ -3,10 +3,10 @@ from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticatedOrReadOnly, AllowAny
 from rest_framework import status
 from accounts.models import CustomUser
-from articles.models import Article, Comment, ArticleLike, CommentLike
+from articles.models import Article, Comment
 from .serializers import ProfileSerializer, ProfileUpdateSerializer
 from articles.serializers import ArticleListSerializer, CommentSerializer
-from .validators import validate_user_data, validate_password_change
+from .validators import validate_user_data
 
 
 # 프로필
@@ -27,7 +27,7 @@ class ProfileView(APIView):
             return Response({'detail': '본인의 프로필만 수정할 수 있습니다.'}, status=status.HTTP_403_FORBIDDEN)
 
         # 데이터 유효성 검사
-        rlt_message = validate_user_data(request.data, request.user)
+        rlt_message = validate_user_data(request.data)
         if rlt_message is not None:
             return Response({"message": rlt_message}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -41,22 +41,6 @@ class ProfileView(APIView):
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class PasswordChangeView(APIView):
-    def post(self, request):
-        old_password = request.data.get("old_password")
-        new_password = request.data.get("new_password")
-
-        rlt_message = validate_password_change(
-            new_password, old_password, request.user)
-        if rlt_message:
-            return Response({"message": rlt_message}, status=status.HTTP_400_BAD_REQUEST)
-
-        # 비밀번호 변경 처리
-        request.user.set_password(new_password)
-        request.user.save()
-        return Response({"message": "비밀번호가 성공적으로 변경되었습니다."}, status=status.HTTP_200_OK)
 
 
 class ArticlesView(APIView):
